Refactored SessionCompleteButton from courses/[slug]/[sessionSlug]
This allows the component to be re-used in the IBA course and the other courses where it has now been plugged in.
